Moved AnnoyanceHandlerProvider to the package "annoyancehandlers"
